feat(web): catalogs architecture refactoring
