+++ apps/web/src/features/catalog-forms/AccountForm/index.ts
@@ -1,0 +1,1 @@
+export { AccountForm } from './AccountForm';
