+++ apps/web/src/features/catalog-forms/ArticleForm/index.ts
@@ -1,0 +1,1 @@
+export { ArticleForm } from './ArticleForm';
