+++ apps/web/src/features/catalog-forms/CounterpartyForm/index.ts
@@ -1,0 +1,1 @@
+export { CounterpartyForm } from './CounterpartyForm';
