+++ apps/web/src/features/catalog-forms/DealForm/index.ts
@@ -1,0 +1,1 @@
+export { DealForm } from './DealForm';
