+++ apps/web/src/features/catalog-forms/DepartmentForm/index.ts
@@ -1,0 +1,1 @@
+export { DepartmentForm } from './DepartmentForm';
