+++ apps/web/src/features/catalog-forms/SalaryForm/index.ts
@@ -1,0 +1,1 @@
+export { SalaryForm } from './SalaryForm';
