+++ apps/web/src/features/catalog-forms/index.ts
@@ -1,0 +1,6 @@
+export { AccountForm } from './AccountForm';
+export { ArticleForm } from './ArticleForm';
+export { DepartmentForm } from './DepartmentForm';
+export { CounterpartyForm } from './CounterpartyForm';
+export { DealForm } from './DealForm';
+export { SalaryForm } from './SalaryForm';
