+++ apps/web/src/pages/catalogs/CounterpartiesPage.tsx
@@ -5,8 +5,6 @@ import { Layout } from '../../shared/ui/Layout';
 import { Card } from '../../shared/ui/Card';
 import { Button } from '../../shared/ui/Button';
 import { Table } from '../../shared/ui/Table';
-import { Input } from '../../shared/ui/Input';
-import { Select } from '../../shared/ui/Select';
 import {
   useGetCounterpartiesQuery,
   useCreateCounterpartyMutation,
@@ -15,6 +13,7 @@ import {
 } from '../../store/api/catalogsApi';
 import type { Counterparty } from '@shared/types/catalogs';
 import { OffCanvas } from '@/shared/ui/OffCanvas';
+import { CounterpartyForm } from '@/features/catalog-forms/index';
 
 export const CounterpartiesPage = () => {
   const [isFormOpen, setIsFormOpen] = useState(false);
@@ -101,67 +100,5 @@ export const CounterpartiesPage = () => {
         />
       </OffCanvas>
     </Layout>
-  );
-};
-
-export const CounterpartyForm = ({
-  counterparty,
-  onClose,
-}: {
-  counterparty: Counterparty | null;
-  onClose: () => void;
-}) => {
-  const [name, setName] = useState(counterparty?.name || '');
-  const [inn, setInn] = useState(counterparty?.inn || '');
-  const [category, setCategory] = useState(counterparty?.category || 'other');
-  const [create, { isLoading: isCreating }] = useCreateCounterpartyMutation();
-  const [update, { isLoading: isUpdating }] = useUpdateCounterpartyMutation();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    try {
-      if (counterparty)
-        await update({
-          id: counterparty.id,
-          data: { name, inn, category },
-        }).unwrap();
-      else await create({ name, inn, category }).unwrap();
-      onClose();
-    } catch (error) {
-      console.error('Failed to save counterparty:', error);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="space-y-4">
-      <Input
-        label="Название"
-        value={name}
-        onChange={(e) => setName(e.target.value)}
-        required
-      />
-      <Input label="ИНН" value={inn} onChange={(e) => setInn(e.target.value)} />
-      <Select
-        label="Категория"
-        value={category}
-        onChange={(e) => setCategory(e.target.value)}
-        options={[
-          { value: 'supplier', label: 'Поставщик' },
-          { value: 'customer', label: 'Клиент' },
-          { value: 'gov', label: 'Гос. орган' },
-          { value: 'employee', label: 'Сотрудник' },
-          { value: 'other', label: 'Другое' },
-        ]}
-        required
-      />
-      <div className="flex gap-4 pt-4">
-        <Button type="submit" disabled={isCreating || isUpdating}>
-          {counterparty ? 'Сохранить' : 'Создать'}
-        </Button>
-        <Button type="button" variant="secondary" onClick={onClose}>
-          Отмена
-        </Button>
-      </div>
-    </form>
   );
 };
